Initial working version of the OpenNMS Drift deployment.
The terraform recipes will generate all the infrastructure using plain EC2 instances.

There is still work to do in order to have telemetry and flow working properly, but the preliminary tests looks promising.

This requires a Minion to work, and a Vagrant recipe is included with this repository.

A fleet for UI servers and Grafana will be added in the future.