Restoring default topic names for the Kafka Producer to have OCE working.
It is not clear the correct place to override the topic names when
configuring OCE. For this reason, I've restored the defaults, and
now I can see that OCE can see the alarms through the `oce:alarms`
command in Karaf.

The JVM settings for Minion and Sentinel have been updated.